requirements.txt
# Core dependencies
google-genai>=0.2.0
ollama>=0.1.0
python-dotenv>=1.0.0

# PostgreSQL
psycopg2-binary>=2.9.9

# UI/CLI
rich>=13.7.0

# Image Processing (for Vision)
Pillow>=10.0.0
opencv-python>=4.8.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
